create engels.db module
