+++ engels/loader.py
@@ -5,9 +5,8 @@ import os
 import re
 import tempfile
 
-from sqlalchemy import create_engine
-
 from .common import get_data_path, get_db_data_path, get_load_manifest
+from .db import get_db_engine
 
 class AutoName(Enum):
     def _generate_next_value_(name, start, count, last_values):
@@ -21,13 +20,6 @@ class FileFormat(AutoName):
     def parse(cls, format_str):
         return cls(format_str.upper())
 
-
-def get_db_engine(user=None, password=None, hostname=None):
-    user = user or os.getenv('DB_USER')
-    password = password or os.getenv('DB_PASSWORD')
-    hostname = hostname or os.getenv('DB_HOSTNAME')
-    engine = create_engine(f'postgresql+psycopg2://{user}:{password}@{hostname}')
-    return engine
 
 CAMEL_TO_SNAKE_RE = re.compile(r'(?<!^)(?=[A-Z])')
